# be/swagger.json
{
    "info": {
        "title": "Taxi Booking API",
        "version": "1.0.0",
        "description": "RESTful API description with Swagger"
    },
    "host": "localhost:3000",
    "swagger": "2.0",
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Authentication",
                    "Profile"
                ],
                "summary": "Get logged in user",
                "description": null,
                "operationId": "profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "User not found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authentication",
                    "Update Profile"
                ],
                "summary": "Get logged in user",
                "description": null,
                "operationId": "updateprofile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string",
                        "value":"Bearer "
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateProfile"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "User not found"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
     },
    "definitions": {
...
            "UpdateProfile": {
                "type": "object",
                "required": [
                    "firstname",
                    "lastname",
                    "mobile_number",
                    "password",
                    "role_type"
                ],
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile_number": {
                        "type": "string"
                    }
                    
                }
            },
...
    }
}